Fixed warnings
git-svn-id: http://javascool.googlecode.com/svn@82 17e3714e-118d-3e7a-01ef-bee73ecf6174

+++ trunk/src/org/javascool/proglets/game/Oval.java
@@ -30,10 +30,10 @@ public class Oval extends Geometry implements Drawable {
     @Override
     public void draw(Graphics g) {
         if (isM_solid()) {
-            g.fillOval((int)getX(), (int)getY(), (int)getWidth(), (int)getHeight());
+            g.fillOval(getX(), getY(), getWidth(), getHeight());
         }
         else {
-            g.drawOval((int)getX(), (int)getY(), (int)getWidth(), (int)getHeight());
+            g.drawOval(getX(), getY(), getWidth(), getHeight());
         }
     }
     private static final Logger LOG = Logger.getLogger(Oval.class.getName());
